test tool checkpoint

+++ perftool/perftool/main.py
@@ -7,16 +7,20 @@ import shutil
 from utils import (
     build_release,
     create_repo_from_commmit,
+    get_cli_repo,
+    get_csv_file,
     init_csv,
     load_bench_function_data,
     sample_and_bench,
     write_into_csv,
 )
 
+TEMPFOLDERNAME: str = "Perftool_Repos_tmp"
+
 app = typer.Typer()
 
 
-def load_commits_from_config(bench_function: str) -> dict:
+def load_commits_from_config(bench_function: str) -> dict[str, str]:
     bench_function_data = load_bench_function_data(bench_function)
     commits = bench_function_data["commits"]
     return commits
@@ -30,7 +34,7 @@ def build_repo(commit: str, tmpfolder: Path):
         print(f"{e}")
         print(f"Skipping build for {commit}...")
         return
-    create_repo_from_commmit(commit, str(commit_folder))
+    create_repo_from_commmit(commit, commit_folder)
     cli_repo = commit_folder / "cli"
     build_release(cli_repo)
 
@@ -52,39 +56,29 @@ def create_symlink_for_repo(commit: str, tmpfolder: Path, bench_function: str):
 
 @app.command()
 def bench(bench_function: str, min_value: int, max_value: int):
-    commits = load_commits_from_config(bench_function)
-    commit_1, commit_2 = list(commits.values())[:2]  # change later
-
+    bench_commits = load_commits_from_config(bench_function)
+    # create build folder if it doesn't exist
     build_folder = Path.cwd() / "build"
     build_folder.mkdir(exist_ok=True)
+    # create bench folder if it doesn't exist
     bench_folder = build_folder / bench_function
-    commit_1_symlink = bench_folder / commit_1
-    commit_2_symlink = bench_folder / commit_2
-    commit_1_folder, commit_2_folder = (
-        commit_1_symlink.resolve(),
-        commit_2_symlink.resolve(),
-    )
-    cli_repo_1, cli_repo_2 = commit_1_folder / "cli", commit_2_folder / "cli"
-
+    bench_folder.mkdir(exist_ok=True)
+    # create data folder if it doesn't exist
     data_folder = Path.cwd() / "data"
     data_folder.mkdir(exist_ok=True)
-    bench_folder = data_folder / bench_function
-    bench_folder.mkdir(exist_ok=True)
-    data_1_csv_file, data_2_csv_file = (
-        bench_folder / f"{commit_1}.csv",
-        bench_folder / f"{commit_2}.csv",
-    )
+    commits = list(commit for (_commit_description, commit) in bench_commits)
+    # initialize the csv files with headers if they does not exist
+    for commit in commits:
+        data_csv_file = bench_folder / f"{commit}.csv"
+        init_csv(data_csv_file, bench_function)
 
-    # initialize the csv files with headers if it does not exist
-    init_csv(data_1_csv_file, bench_function)
-    init_csv(data_2_csv_file, bench_function)
     num_samples = 0
     while True:
         try:
-            (cli_repo, data_csv_file) = random.choice(
-                [(cli_repo_1, data_1_csv_file), (cli_repo_2, data_2_csv_file)]
-            )
+            commit = random.choice(commits)
+            cli_repo = get_cli_repo(commit, bench_function)
             data = sample_and_bench(cli_repo, bench_function, min_value, max_value)
+            data_csv_file = get_csv_file(commit, bench_function)
             write_into_csv(data, data_csv_file)
             num_samples += 1
         except KeyboardInterrupt:
@@ -95,27 +89,40 @@ def bench(bench_function: str, min_value: int, max_value: int):
 
 @app.command()
 def build(bench_function: str):
-    commits = load_commits_from_config(bench_function)
-    commit_1, commit_2 = list(commits.values())[:2]  # change later
-    tmpfolder = Path(tempfile.gettempdir()) / "Perftool_Repos_tmp"
+    bench_commits = load_commits_from_config(bench_function)
+    tmpfolder = Path(tempfile.gettempdir()) / TEMPFOLDERNAME
     tmpfolder.mkdir(exist_ok=True)
-    build_repo(commit_1, tmpfolder)
-    build_repo(commit_2, tmpfolder)
-    create_symlink_for_repo(commit_1, tmpfolder, bench_function)
-    create_symlink_for_repo(commit_2, tmpfolder, bench_function)
+    for commit_description, commit in bench_commits.items():
+        build_repo(commit, tmpfolder)
+        create_symlink_for_repo(commit, tmpfolder, bench_function)
     print(f"Bench {bench_function} built succesfully.")
 
 
 @app.command()
 def clean(bench_function: str):
     bench_commits_folder = Path.cwd() / "build" / bench_function
+    if not bench_commits_folder.exists():
+        print("No bench commits found")
+        return
     for commit_symlink in bench_commits_folder.iterdir():
         commit_folder = commit_symlink.resolve()
         if commit_folder.is_dir():
             shutil.rmtree(commit_folder)
         commit_symlink.unlink()
     bench_commits_folder.rmdir()
-    print("Cleaned successfully")
+    print("Repos cleaned successfully")
+
+
+@app.command()
+def cleancsv(bench_function: str):
+    bench_commits_folder = Path.cwd() / "data" / bench_function
+    if not bench_commits_folder.exists():
+        print("No bench csv files found")
+        return
+    for commit_csv_file in bench_commits_folder.iterdir():
+        commit_csv_file.unlink()
+    bench_commits_folder.rmdir()
+    print("Csv files sleaned successfully")
 
 
 if __name__ == "__main__":
